* subversion/libsvn_ra_neon/props.c   (svn_ra_neon__get_props_resource): One more bitness mismatch fix.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1174163 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_ra_neon/props.c
@@ -571,7 +571,7 @@ svn_error_t * svn_ra_neon__get_props_resource(svn_ra_neon__resource_t **rsrc,
 {
   apr_hash_t *props;
   char * url_path = apr_pstrdup(pool, url);
-  int len = strlen(url);
+  size_t len = strlen(url);
   /* Clean up any trailing slashes. */
   if (len > 1 && url[len - 1] == '/')
       url_path[len - 1] = '\0';
